- разделил клиентов на логические части
