+++ src/main/java/ru/yandex/practicum/clients/AuthClient.java
@@ -7,9 +7,7 @@ import ru.yandex.practicum.models.auth.UserRegisterRequest;
 
 import static io.restassured.RestAssured.given;
 
-public class Client {
-
-    protected static final String QA_STELLARBURGERS_SERVICE = "https://stellarburgers.nomoreparties.site";
+public class AuthClient extends ClientBase {
 
     private static final String API_AUTH_REGISTER = "/api/auth/register";
     private static final String API_AUTH_LOGIN = "/api/auth/login";
@@ -17,9 +15,6 @@ public class Client {
     private static final String API_AUTH_TOKEN = "/api/auth/token";
     private static final String API_AUTH_USER = "/api/auth/user";
 
-    private static final String API_INGREDIENTS = "/api/ingredients";
-    private static final String API_ORDERS = "/api/orders";
-    private static final String API_ORDERS_ALL = "/api/orders/all";
     private static final String API_PASSWORD_RESET = "/api/password-reset";
 
     public Response registerUser(UserRegisterRequest userRegisterRequest) {
@@ -49,7 +44,7 @@ public class Client {
                 .post(API_AUTH_LOGIN);
     }
 
-    public Response updateUser(String accessToken, UserInfoUpdateRequest userInfoUpdateRequest){
+    public Response updateUser(String accessToken, UserInfoUpdateRequest userInfoUpdateRequest) {
         return given()
                 .baseUri(QA_STELLARBURGERS_SERVICE)
                 .header("Content-type", "application/json")
@@ -57,13 +52,5 @@ public class Client {
                 .body(userInfoUpdateRequest)
                 .when()
                 .patch(API_AUTH_USER);
-    }
-
-    public Response getIngredients() {
-        return given()
-                .baseUri(QA_STELLARBURGERS_SERVICE)
-                .header("Content-type", "application/json")
-                .when()
-                .get(API_INGREDIENTS);
     }
 }
