+++ src/main/java/ru/yandex/practicum/clients/ClientBase.java
@@ -1,0 +1,6 @@
+package ru.yandex.practicum.clients;
+
+public class ClientBase {
+
+    protected static final String QA_STELLARBURGERS_SERVICE = "https://stellarburgers.nomoreparties.site";
+}
