+++ src/main/java/ru/yandex/practicum/clients/OrderClient.java
@@ -1,0 +1,31 @@
+package ru.yandex.practicum.clients;
+
+import io.restassured.response.Response;
+import ru.yandex.practicum.models.orders.CreateOrderRequest;
+
+import static io.restassured.RestAssured.given;
+
+public class OrderClient extends ClientBase{
+
+    private static final String API_INGREDIENTS = "/api/ingredients";
+    private static final String API_ORDERS = "/api/orders";
+    private static final String API_ORDERS_ALL = "/api/orders/all";
+
+    public Response createOrder(String accessToken, CreateOrderRequest createOrderRequest) {
+        return given()
+                .baseUri(QA_STELLARBURGERS_SERVICE)
+                .header("Content-type", "application/json")
+                .header("Authorization", accessToken)
+                .body(createOrderRequest)
+                .when()
+                .post(API_ORDERS);
+    }
+
+    public Response getIngredients() {
+        return given()
+                .baseUri(QA_STELLARBURGERS_SERVICE)
+                .header("Content-type", "application/json")
+                .when()
+                .get(API_INGREDIENTS);
+    }
+}
