+++ src/test/java/ru/yandex/practicum/auth/UpdateUserTests.java
@@ -18,7 +18,7 @@ public class UpdateUserTests extends LoginUserTestBase {
     @DisplayName("Проверка обновления информации пользователя")
     @Description("Позитивная проверка возможности обновить данные пользователя")
     public void userInfoCanBeUpdatedTest(){
-        Response updatedUserInfoResponse = this.client.updateUser(this.registeredUser.getAccessToken(), this.updatedUserInfo);
+        Response updatedUserInfoResponse = this.authClient.updateUser(this.registeredUser.getAccessToken(), this.updatedUserInfo);
         assertEquals("Неверный статус-код", HttpStatus.SC_OK, updatedUserInfoResponse.statusCode());
 
         UserInfoResponse userInfoResponse = updatedUserInfoResponse.as(UserInfoResponse.class);
@@ -31,7 +31,7 @@ public class UpdateUserTests extends LoginUserTestBase {
     @DisplayName("Проверка невозможности обновления информации пользователя")
     @Description("Проверка невозможности обновить данные пользователя без авторизации")
     public void unableToUpdateUserInfoWithoutAuthorizationTest(){
-        Response updatedUserInfoResponse = this.client.updateUser("", this.updatedUserInfo);
+        Response updatedUserInfoResponse = this.authClient.updateUser("", this.updatedUserInfo);
         assertEquals("Неверный статус-код", HttpStatus.SC_UNAUTHORIZED, updatedUserInfoResponse.statusCode());
 
         ErrorResponse errorResponse = updatedUserInfoResponse.as(ErrorResponse.class);
